pyproject.toml
[project]
name = "ymca"
version = "0.1.0"
description = "YNAB FX transaction conversion CLI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyYAML>=6.0.2",
    "ynab>=4.0.0",
]

[project.scripts]
ymca = "ymca:main"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.14.1",
    "types-PyYAML>=6.0.12.20250915",
]

[tool.mypy]
python_version = "3.13"
strict = true

[[tool.mypy.overrides]]
module = ["ynab", "ynab.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"
